Adds PawnFactory and manages to squash every compile time error
