+++ app/src/main/java/com/michaeljahns/namespace/ScenarioPageAdapter.kt
@@ -9,21 +9,22 @@ import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
 import com.michaeljahns.namespace.grammy.Scenario
 
-class ViewPageAdapter(private var scenarios: MutableList<Scenario>) : RecyclerView.Adapter<ViewPageAdapter.ViewPager2Holder>() {
+class ScenarioPageAdapter(private var scenarios: MutableList<Scenario>) : RecyclerView.Adapter<ScenarioPageAdapter.ViewPager2Holder>() {
 
 
-    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): ViewPageAdapter.ViewPager2Holder {
+    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): ScenarioPageAdapter.ViewPager2Holder {
         val view: View = LayoutInflater.from(parent.context).inflate(R.layout.scenario_row, parent, false)
 //        Explicating declaring the below resolved some issue, i cant remember now which
         view.layoutParams = ViewGroup.LayoutParams(ViewGroup.LayoutParams.MATCH_PARENT, ViewGroup.LayoutParams.MATCH_PARENT)
         return ViewPager2Holder(view)
     }
 
-    override fun onBindViewHolder(holder: ViewPageAdapter.ViewPager2Holder, position: Int) {
+    override fun onBindViewHolder(holder: ScenarioPageAdapter.ViewPager2Holder, position: Int) {
         val scenario = scenarios[position]
         holder.scenarioLocation.text = scenario.location.toString()
         val childLayoutManager = LinearLayoutManager(holder.scenarioPawnRecycler.context, RecyclerView.HORIZONTAL, false)
-        holder.scenarioPawnRecycler.
+        holder.scenarioPawnRecycler.layoutManager = childLayoutManager
+        holder.scenarioPawnRecycler.adapter = PawnRecyclerAdapter(scenario.pawnList)
     }
 
     override fun getItemCount(): Int {
@@ -42,24 +43,3 @@ class ViewPageAdapter(private var scenarios: MutableList<Scenario>) : RecyclerVi
         }
     }
 }
-//
-//    class ParentAdapter(private val parents: List<ParentModel>) : RecyclerView.Adapter<ParentAdapter.ViewHolder>() {
-//        private val viewPool = RecyclerView.RecycledViewPool()
-//
-//        override fun onBindViewHolder(holder: ViewHolder,
-//                                      position: Int) {
-//            val parent = parents[position]
-//            holder.recyclerView.apply {
-//                layoutManager = childLayoutManager
-//                adapter = ChildAdapter(parent.children)
-//                setRecycledViewPool(viewPool)
-//            }
-//
-//        }
-//
-//        inner class ViewHolder(itemView: View) : RecyclerView.ViewHolder(itemView) {
-//            val recyclerView: RecyclerView = itemView.rv_child
-//            val textView: TextView = itemView.textView
-//        }
-//    }
-//}
